Moved PROJECTION_*** constants from /scene to /core/shape
