+++ src/core/shape/constants.js
@@ -1,0 +1,13 @@
+/**
+ * A perspective camera projection where the frustum shape is essentially pyramidal.
+ *
+ * @type {number}
+ */
+export const PROJECTION_PERSPECTIVE = 0;
+
+ /**
+  * An orthographic camera projection where the frustum shape is essentially a cuboid.
+  *
+  * @type {number}
+  */
+export const PROJECTION_ORTHOGRAPHIC = 1;
